Implement fill for antique versions of libcheck.
The version of libcheck used by TravisCI's default linux container
belongs in a museum, and does not have the ck_assert_ptr_nonnull()
macro.

The nostalgia is bracing.

In order to allow Travis' creaky builds to proceed unimpeded by such
newfangled modern contrivances, let's just define a macro and be on our
way.