Update shimmy2 hash implementation.
I was not happy about information loss -- once we have processed 32
characters from the source string, the bits that were shifted off the
left of the hash were effectively gone.  So two different strings that
happened to end in the same 32 characters would yield the same hash.

I modified the algorithm to fold the bit that is shifted off into the
incoming byte.  I also changed my mind about multiplying by the N+1 byte
and decided to multiply by the N-1 byte instead.